+++ tests/testUtils.ts
@@ -1,4 +1,4 @@
-export const isValidUrl = (string) => {
+export const isValidUrl = (string: string) => {
   try {
     new URL(string);
     return true;
@@ -7,7 +7,7 @@ export const isValidUrl = (string) => {
   }
 };
 
-export const isValidEmail = (string) => {
+export const isValidEmail = (string: string) => {
   const emailRegExp =
     /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|.(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;